♻️ refactor: 도메인 hook 폴더 이동 - /application/hooks/domain -> /features/common/hooks
