--- src/application/hooks/domain/channel/index.ts
@@ -1,1 +1,0 @@
-export * from "./useChannelIO";
