--- src/application/hooks/domain/collection/index.ts
@@ -1,1 +1,0 @@
-export * from "./useCollection";
