+++ src/application/hooks/domain/index.ts
@@ -1,6 +1,1 @@
 export * from "./auth";
-export * from "./channel";
-export * from "./collection";
-export * from "./meme";
-export * from "./search";
-export * from "./share";
