--- src/application/hooks/domain/meme/index.ts
@@ -1,1 +1,0 @@
-export * from "./useMoveMemeDetail";
