+++ src/features/common/components/MemeActionSheet.tsx
@@ -1,12 +1,14 @@
 import { Suspense } from "react";
 import tw from "twin.macro";
 
-import { useAuth, useCollection } from "@/application/hooks";
+import { useAuth } from "@/application/hooks";
 import { ActionSheet } from "@/common/components/ActionSheet";
 import { useOverlay } from "@/common/hooks";
 import { channelUrl } from "@/common/utils";
 import { MemeShareModal } from "@/features/memes/components";
 import type { Meme } from "@/types";
+
+import { useCollection } from "../hooks";
 
 interface Props {
   meme: Meme;
