+++ src/features/memes/components/MemeCTAList.tsx
@@ -1,9 +1,10 @@
 import { Suspense } from "react";
 
-import { useAuth, useCollection } from "@/application/hooks";
+import { useAuth } from "@/application/hooks";
 import { Button } from "@/common/components/Button";
 import { Icon } from "@/common/components/Icon";
 import { useOverlay } from "@/common/hooks";
+import { useCollection } from "@/features/common";
 
 import { MemeShareModal } from "./MemeShareModal";
