+++ src/features/search/components/SearchRecent.tsx
@@ -1,10 +1,10 @@
 import { useRouter } from "next/router";
 
-import type { RecentSearch } from "@/application/hooks";
-import { isTagType } from "@/application/hooks";
 import { Icon } from "@/common/components/Icon";
 import { PATH } from "@/common/utils";
 
+import type { RecentSearch } from "../hooks";
+import { isTagType } from "../hooks";
 import { SearchItem } from "./SearchItem";
 
 interface Props {
