+++ src/pages/memes/[id].tsx
@@ -2,12 +2,13 @@ import { dehydrate, QueryClient } from "@tanstack/react-query";
 import type { GetStaticPaths, GetStaticProps, NextPage } from "next";
 import { Suspense } from "react";
 
-import { fetchMemeDetailById, fetchMemeTagsById, useMoveMemeDetail } from "@/application/hooks";
+import { fetchMemeDetailById, fetchMemeTagsById } from "@/application/hooks";
 import { ExplorePageNavigation } from "@/common/components/Navigation";
 import { NextSeo } from "@/common/components/NextSeo";
 import { MemeListSkeleton, Skeleton } from "@/common/components/Skeleton";
 import { SSRSuspense } from "@/common/components/Suspense";
 import { SITE_NAME } from "@/common/utils";
+import { useMoveMemeDetail } from "@/features/common";
 import {
   MemeCTAList,
   MemeDetail,
