+++ src/pages/mypage/index.tsx
@@ -1,13 +1,13 @@
 import Link from "next/link";
 import { css } from "twin.macro";
 
-import { useAuth, useChannelIO } from "@/application/hooks";
+import { useAuth } from "@/application/hooks";
 import { Icon } from "@/common/components/Icon";
 import { MyPageNavigation } from "@/common/components/Navigation";
 import { Photo } from "@/common/components/Photo";
 import { MemeListSkeleton, Skeleton } from "@/common/components/Skeleton";
 import { SSRSuspense } from "@/common/components/Suspense";
-import { withAuth } from "@/features/common";
+import { useChannelIO, withAuth } from "@/features/common";
 import { SummarizedCollection } from "@/features/mypage/components";
 
 const MyPage = () => {
