+++ src/pages/search/index.tsx
@@ -2,7 +2,6 @@ import type { NextPage } from "next";
 import { useRouter } from "next/router";
 import { Suspense } from "react";
 
-import { useRecentSearch } from "@/application/hooks";
 import { SearchPageNavigation } from "@/common/components/Navigation";
 import type { NextSeoProps } from "@/common/components/NextSeo";
 import { NextSeo } from "@/common/components/NextSeo";
@@ -16,6 +15,7 @@ import {
   SearchResultList,
   SkeletonTagList,
 } from "@/features/search/components";
+import { useRecentSearch } from "@/features/search/hooks";
 
 /**
  * FIX
